+++ src/styleguidist/ExampleForm.js
@@ -1,6 +1,7 @@
 import React from 'react';
 import T from 'prop-types';
-import { reduxForm } from 'redux-form';
+import { reduxForm, getFormValues } from 'redux-form';
+import { connect } from 'react-redux';
 
 import { Button } from '../lib';
 
@@ -13,22 +14,54 @@ export class ExampleForm extends React.Component {
   static propTypes = {
     handleSubmit: T.func.isRequired,
     reset: T.func.isRequired,
+    formValues: T.object,
     children: T.node.isRequired,
+    customButton: T.object,
   }
+
+  static defaultProps = {
+    formValues: {},
+    customButton: null,
+  }
+
+  state = { showValues: false }
 
   onSubmit = (values) => {
     alert(JSON.stringify(values, null, 2));
   }
 
+  toggleValues = () => {
+    this.setState(state => ({ showValues: !state.showValues }));
+  }
+
+  clickCustomButton = () => {
+    const { customButton, formValues } = this.props;
+    customButton.onClick(formValues);
+  }
+
   render() {
-    const { handleSubmit, children, reset } = this.props;
+    const {
+      handleSubmit, children, reset, formValues, customButton,
+    } = this.props;
+    const { showValues } = this.state;
+
+    const valuesText = showValues ? 'Hide Values' : 'Show Values';
 
     return (
-      <form onSubmit={handleSubmit(this.onSubmit)}>
-        {children}
-        <Button type={'submit'} text={'Submit'} variant={'primary'} />
-        <Button text={'Reset'} onClick={reset} variant={'default'} />
-      </form>
+      <div>
+        <form onSubmit={handleSubmit(this.onSubmit)}>
+          {children}
+          <Button type={'submit'} text={'Submit'} variant={'primary'} />
+          <Button text={'Reset'} onClick={reset} variant={'default'} />
+          <Button text={valuesText} onClick={this.toggleValues} variant={'standout'} className={'float-right'} />
+          {customButton && <Button {...customButton} className={'float-right'} onClick={this.clickCustomButton} />}
+        </form>
+
+        {
+          showValues
+          && <pre className={'m-t-sm'}>{JSON.stringify(formValues, null, 2)}</pre>
+        }
+      </div>
     );
   }
 }
@@ -36,5 +69,9 @@ export class ExampleForm extends React.Component {
 const withReduxForm = reduxForm();
 const WrappedForm = withReduxForm(ExampleForm);
 
+const mapStateToProps = (state, props) => ({
+  formValues: getFormValues(props.form)(state),
+});
+
 // pass down props, allowing to dynamically generate form name...
-export default props => <WrappedForm {...props} />;
+export default connect(mapStateToProps)(WrappedForm);